Split different FOCDMDs into separate plots. Fixed circular import.
